requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jwt==4.1.0
PyJWT==2.8.0
openai==1.3.7
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
